requirements.txt
pytest==3.7.3
request=2.21.0
pytest-allure-adaptor=1.7.10
